+++ src/client/components/TablePaginated/TablePaginated.tsx
@@ -1,6 +1,7 @@
 import './TablePaginated.scss'
 import React from 'react'
 
+import { useTablePaginatedCount } from 'client/store/ui/tablePaginated'
 import DataGrid from 'client/components/DataGrid'
 
 import { useFetchData } from './hooks/useFetchData'
@@ -12,9 +13,19 @@ import { Props } from './types'
 
 type TablePaginatedProps<Datum extends object> = Props<Datum> & { className?: string; header?: boolean }
 const TablePaginated = <Datum extends object>(props: TablePaginatedProps<Datum>) => {
-  const { className, columns, header, path, limit } = props
+  const { className, columns, header, path, limit, EmptyListComponent } = props
 
   useFetchData({ path, limit })
+
+  const count = useTablePaginatedCount(path)
+
+  if (count?.total === 0) {
+    return (
+      <div className={className}>
+        <EmptyListComponent />
+      </div>
+    )
+  }
 
   return (
     <div className={className}>
