+++ src/client/components/TablePaginated/types.ts
@@ -8,6 +8,7 @@ export type Column<Datum> = {
 }
 
 export type Props<Datum> = {
+  EmptyListComponent?: React.FC
   columns: Array<Column<Datum>>
   path: string
   limit?: number
