+++ src/client/pages/CountryHome/FraHome/RecentActivity/EmptyActivities/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EmptyActivities'
